requirements.txt
ffmpeg        1.4
load-dotenv   0.1.0
pip           24.0
python-dotenv 1.0.1
